fixed code reviews bugs
