+++ src/app/page.tsx
@@ -1,8 +1,6 @@
 'use client';
 
 import { useState, ChangeEvent } from 'react';
-
-import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
 import toast, { Toaster } from 'react-hot-toast';
 import { FaHome } from 'react-icons/fa';
@@ -10,11 +8,12 @@ import { IoIosAddCircleOutline } from 'react-icons/io';
 import { CgCalendar } from 'react-icons/cg';
 import { RiDeleteBin6Line } from 'react-icons/ri';
 import Drawer from './components/modals/drawer';
+import DateSelector from './components/DateSelector';
 
 
 type Task = {
   description: string;
-  status: 'completed' | 'notCompleted';
+  status: boolean;
   date: string;
 };
 
@@ -24,8 +23,6 @@ export default function Home() {
   const [taskDate, setTaskDate] = useState('');
   const [isSidebarOpen, setIsSidebarOpen] = useState(false);
   const [selectedTask, setSelectedTask] = useState<Task | null>(null);
-  const [isCalendarModalOpen, setIsCalendarModalOpen] = useState(false);
-  const [selectedDate, setSelectedDate] = useState<Date | null>(null);
   const [showDateMenu, setShowDateMenu] = useState(false);
 
   const handleAddTask = () => {
@@ -40,7 +37,7 @@ export default function Home() {
    
     const newTask: Task = {
       description: task,
-      status: 'notCompleted',
+      status: false,
       date: taskDate,
     };
     setTasks([newTask, ...tasks]);
@@ -50,13 +47,12 @@ export default function Home() {
 
   const tasksUpdated = (index: number) => {
     const updated = [...tasks];
-    updated[index].status =
-      updated[index].status === 'completed' ? 'notCompleted' : 'completed';
+    updated[index].status = !updated[index].status;
     setTasks(updated);
   };
 
   const deleteTask = (index: number) => {
-    if (tasks[index].status === 'completed') {
+    if (tasks[index].status) {
       const updated = tasks.filter((_, i) => i !== index);
       setTasks(updated);
     }
@@ -72,8 +68,8 @@ export default function Home() {
   };
   const sortedTasks =tasks
   .sort((a, b) => {
-    if (a.status === 'completed' && b.status !== 'completed') return 1;
-    if (a.status !== 'completed' && b.status === 'completed') return -1;
+    if (a.status  && !b.status ) return 1;
+    if (!a.status  && b.status ) return -1;
     return 0;
   })
 
@@ -99,13 +95,13 @@ export default function Home() {
                 <div>
                 <input
                   type="checkbox"
-                  checked={item.status === 'completed'}
+                  checked={item.status}
                   onClick={(e) => e.stopPropagation()}
                   onChange={() => tasksUpdated(index)}
                 />
                 </div>
                 <div>
-                  <p className={item.status === 'completed' ? 'line-through text-gray-500' : ''}>
+                  <p className={item.status  ? 'line-through text-gray-500' : ''}>
                     {item.description}
                   </p>
                   <p className="text-xs text-gray-400">{item.date}</p>
@@ -144,67 +140,20 @@ export default function Home() {
           </button>
           </div>
           {showDateMenu && (
-            <select
-              className="text-sm white border rounded px-3 py-1 absolute bottom-1 right-0 bg-white"
-              onChange={(e) => {
-                const option = e.target.value;
-                const today = new Date();
-                const format = (d: Date) => d.toISOString().split('T')[0];
-
-                if (option === 'today') {
-                  setTaskDate(format(today));
+            
+              <DateSelector
+                onDateSelect={(date) => {
+                  setTaskDate(date);
                   setShowDateMenu(false);
-                } else if (option === 'tomorrow') {
-                  today.setDate(today.getDate() + 1);
-                  setTaskDate(format(today));
-                  setShowDateMenu(false);
-                } else if (option === 'nextWeek') {
-                  today.setDate(today.getDate() + 7);
-                  setTaskDate(format(today));
-                  setShowDateMenu(false);
-                } else if (option === 'custom') {
-                  setIsCalendarModalOpen(true);
-                  setShowDateMenu(false);
-                }
-              }}
-            >
-             <option value="">Select Date</option> 
-              <option value="today">Today</option>
-              <option value="tomorrow">Tomorrow</option>
-              <option value="nextWeek">Next Week</option>
-              <option value="custom">Pick a date</option>
-            </select>
-          )}
+                }}
+                onClose={() => setShowDateMenu(false)}
+              />
+            )}
+            
+          
         </div>
       </div>
-
-      
-      {isCalendarModalOpen && (
-        <div className="fixed inset-0 flex justify-end items-center bg-transparent bg-opacity-40 z-50 px-[50px]">
-          <div className="bg-white p-6 rounded shadow-md z-50">
-            <h2 className="text-lg font-semibold mb-4">Select a date</h2>
-            <DatePicker
-              selected={selectedDate}
-              onChange={(date: Date | null) => {
-                if (date) {
-                  setSelectedDate(date);
-                  setTaskDate(date.toISOString().split('T')[0]);
-                  setIsCalendarModalOpen(false);
-                }
-              }}
-              inline
-            />
-            <button
-              onClick={() => setIsCalendarModalOpen(false)}
-              className="mt-4 text-red-500"
-            >
-              Cancel
-            </button>
-          </div>
-        </div>
-      )}
-
-      
+    
 <Drawer
   isOpen={isSidebarOpen}
   onClose={() => setIsSidebarOpen(false)}
